+++ src/services/configuration/dto/index.ts
@@ -1,0 +1,2 @@
+export * from "@/services/configuration/dto/schoolConfiguration.dto";
+export * from "@/services/configuration/dto/configurationStatus.dto";