+++ src/services/configuration/index.ts
@@ -1,0 +1,2 @@
+export * from "@/services/configuration/configuration.service";
+export * from "@/services/configuration/dto";